% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAbundanceMapData.R
\name{getAbundanceMapData}
\alias{getAbundanceMapData}
\title{Get Abundance Map Data}
\usage{
getAbundanceMapData(collections, spatial, interval, spatial_features = NULL)
}
\arguments{
\item{collections}{A dataframe containing collection data retrieved from getArthroCollections()}

\item{spatial}{A dataframe of spatial features retrieved from getSpatialFeatures()}

\item{interval}{Calculation interval for abundance, accepts “collection_date”,“Biweek”,“Week”, and “Month"}

\item{spatial_features}{A character vector of spatial feature names to filter by, if NULL, all will be used}
}
\value{
A dataframe of abundance calculations grouped by the select spatial regions
}
\description{
`getAbundanceMapData()` for mosquito data only, function utilizes spatial feature shape data to group abundance calculations.
}
