% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgency.R
\name{getAgency}
\alias{getAgency}
\title{Get Agency data}
\usage{
getAgency(token)
}
\arguments{
\item{token}{Access token retrieved from `getToken()`}
}
\value{
An sf object containing agency data with geometry
}
\description{
Retrieves VectorSurv agency data with shape coordinates converted to sf geometry
}
\examples{
\dontrun{
token <- getToken()
agency <- getAgency(token)
}
}
