% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPools.R
\name{getPools}
\alias{getPools}
\title{Get Pools data}
\usage{
getPools(token, start_year, end_year)
}
\arguments{
\item{token}{access token retrived from getToken()}

\item{start_year}{Beginning of year range}

\item{end_year}{End of year range}
}
\value{
Dataframe of pools data
}
\description{
Retrieves VectorSurv pools data for desired year range
}
\examples{
\dontrun{
token = getToken()
getPools(token, 2020, 2021)}
}
\keyword{pools}
