% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegions.R
\name{getRegions}
\alias{getRegions}
\title{Get region data}
\usage{
getRegions(token)
}
\arguments{
\item{token}{A valid access token returned from `getToken()`}
}
\value{
A dataframe of region data, used internally to merge spatial information to collections
}
\description{
`getSites()` obtains site data for authorized VectorSurv Gateway accounts.
}
