% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSites.R
\name{getSites}
\alias{getSites}
\title{Get sites data}
\usage{
getSites(token)
}
\arguments{
\item{token}{A valid access token returned from `getToken()`}
}
\value{
A dataframe of site data
}
\description{
`getSites()` obtains site data for authorized VectorSurv Gateway accounts.
}
\examples{
\dontrun{
token = getToken()
sites = getSites(token)}
}
