% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialFeatures.R
\name{getSpatialFeatures}
\alias{getSpatialFeatures}
\title{Get Spatial data}
\usage{
getSpatialFeatures(token, agency_ids = NULL)
}
\arguments{
\item{token}{A valid access token returned from `getToken()`}

\item{agency_ids}{Filter on agency id, default to NULL for all available agencies,otherwise provide a vector of agency ids}
}
\value{
A dataframe of region data, used internally to merge spatial information to collections
}
\description{
`getSpatialFeatures()` obtains spatial data for authorized VectorSurv Gateway accounts. Returns a list of all spatial features tied to an account.
}
\examples{
/dontrun{token=getToken()
spatial = getSpatialFeatures(token)}
}
