% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToken.R
\name{getToken}
\alias{getToken}
\title{Get authentication token}
\usage{
getToken()
}
\value{
User token
}
\description{
getToken() returns a token needed to run getArthroCollections() and getPools(). Prints agencies associated with account credentials. The function prompts users for a VectorSurv account credentials.
}
\examples{
 \dontrun{token = getToken()}
}
\keyword{authentication}
