% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAbundanceAnomaly.R
\name{plotAbundanceAnomaly}
\alias{plotAbundanceAnomaly}
\title{Plot abundance anomaly rate}
\usage{
plotAbundanceAnomaly(AbunAnomOutput)
}
\arguments{
\item{AbunAnomOutput}{Output from returned `getAbundanceAnomaly()`}
}
\value{
ggplot object
}
\description{
plotAbundanceAnomaly() plots the output returned from `getInfectionRate()` with confidence intervals using ggplot. Note the output must not be in wide format.
}
\examples{
AbunAnomOutput = getAbundanceAnomaly(sample_collections,"Biweek",
target_year=2020, trap = "CO2", separate_by="species")
plotAbundanceAnomaly(AbunAnomOutput)
}
