% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVectorIndex.R
\name{plotVectorIndex}
\alias{plotVectorIndex}
\title{Plot Vector Index}
\usage{
plotVectorIndex(VIOutput, year)
}
\arguments{
\item{VIOutput}{Output from returned `getVectorIndex()`}

\item{year}{Year to plot vector index on}
}
\value{
ggplot object
}
\description{
plotVectorIndex() plots the output returned from `getVectorIndex()` with confidence intervals using ggplot
}
\examples{
VI = getVectorIndex(sample_collections,sample_pools,
                      interval = "Week",
                      target_disease = "WNV",
                      pt_estimate = "mle", species_list = c("Cx pipiens"),
                      trap_list = c("CO2","GRVD") )
plotVectorIndex(VIOutput = VI, year = 2017)
}
