% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_collections}
\alias{sample_collections}
\title{Sample Arthropod Collections Data}
\format{
A data frame with 200 rows and 10 variables:
\describe{
  \item{\code{collection_id}}{double Collection identification number}
  \item{\code{collection_date}}{character The date the trap was picked up for collection}
  \item{\code{num_trap}}{integer The number of unique traps in operation at one site}
  \item{\code{surv_year}}{double Surveillance year of collection}
  \item{\code{trap_nights}}{integer The number of nights a trap was in the field}
  \item{\code{trap_problem_bit}}{logical If these was an issue with the trap}
  \item{\code{num_count}}{integer Number of arthropods present in collection}
  \item{\code{sex_type}}{character Sex of collected arthropods}
  \item{\code{species_display_name}}{character Species name of collected arthropods}
  \item{\code{trap_acronym}}{character The acronym of the trap placed in the field}
}
}
\source{
\url{https://vectorsurv.org/}
}
\usage{
sample_collections
}
\description{
Sample Arthropod Collections data imitates the essential components of real mosquito collections data
}
\keyword{datasets}
