% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_spatial}
\alias{sample_spatial}
\title{Sample Spatial Data}
\format{
A data frame with 3 rows and 4 variables:
\describe{
 \item{\code{agency}}{character Agency name}
 \item{\code{agency_id}}{interger Unique agency id number}
\item{\code{id}}{interger ID of spatial feature}
  \item{\code{name}}{character Name of spatial feature}
  \item{\code{geometry}}{multipolygon shape geometry spatal feature}
}
}
\source{
\url{https://vectorsurv.org/}
}
\usage{
sample_spatial
}
\description{
Sample Spatial data imitates spatial feature data
}
\keyword{datasets}
