% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfectionRate.R
\name{getInfectionRate}
\alias{getInfectionRate}
\title{Get Infection Rate}
\usage{
getInfectionRate(
  pools,
  interval,
  target_disease,
  pt_estimate,
  scale = 1000,
  species_list = NULL,
  trap_list = NULL
)
}
\arguments{
\item{pools}{Pools data retrieved from `getPools()`}

\item{interval}{Calculation interval for infection rate, accepts “collection_date”,“Biweek”,“Week”, and “Month}

\item{target_disease}{The disease to calculate infection rate for–i.e. “WNV”. Disease acronyms are the accepted input. To see a list of disease acronyms, run `unique(pools$target_acronym)`}

\item{pt_estimate}{The estimation type for infection rate. Options include: “mle”,“bc-mle”, “mir”}

\item{scale}{Constant to multiply infection rate by}

\item{species_list}{Species filter for calculating infection rate species_display_name is the accepted notation. To see a list of species present in your data run `unique(pools$species_display_name)`. If species is unspecified, the default `NULL` will return data for all species in data.}

\item{trap_list}{Trap filter for calculating infection rate. Trap_acronym is the is the accepted notation. Run `unique(pools$trap_acronym)` to see trap types present in your data. If trap_list is unspecified, the default `NULL` will return data for all trap types.}
}
\value{
Infection rate calculation
}
\description{
`getInfectionRate()` requires at least five years prior to the target_year of arthro collections data to calculate for the specified parameters. The function uses the methods of the Gateway Abundance Anomaly calculator, and will not work if there is fewer than five years of data present.
}
\examples{
getInfectionRate(sample_pools,
                 interval = "Biweek",
                 target_disease = "WNV",
                 pt_estimate = "mle",
                 scale = 1000,
                 species_list = list("Cx pipiens"),
                 trap_list = list("CO2"))
}
\keyword{infection}
\keyword{pools}
\keyword{rate}
