% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPools.R
\name{getPools}
\alias{getPools}
\title{Get Pools data}
\usage{
getPools(token, start_year, end_year, arthropod)
}
\arguments{
\item{token}{access token retrived from `getToken()`}

\item{start_year}{Beginning of year range}

\item{end_year}{End of year range}

\item{arthropod}{Specify arthropod type from: 'mosquito', 'tick', 'nontick'}
}
\value{
Dataframe of pools data
}
\description{
Retrieves VectorSurv pools data for desired year range
}
\examples{
\dontrun{
token = getToken()
getPools(token, start_year = 2020, end_year = 2021, arthropod = 'tick')}
}
\keyword{pools}
