% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSites.R
\name{getSites}
\alias{getSites}
\title{Get sites data}
\usage{
getSites(token, agency_ids = NULL)
}
\arguments{
\item{token}{A valid access token returned from `getToken()`}

\item{agency_ids}{Filter on agency id, default to NULL for all available agencies,otherwise provide a vector of agency ids, such as `agency_ids = c(55,56)`}
}
\value{
A dataframe of site data
}
\description{
`getSites()` obtains site data for authorized VectorSurv Gateway accounts.
}
\examples{
\dontrun{
token = getToken()
sites = getSites(token)}
}
