% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialFeatures.R
\name{getSpatialFeatures}
\alias{getSpatialFeatures}
\title{Get Spatial Features data}
\usage{
getSpatialFeatures(token, agency_ids = NULL)
}
\arguments{
\item{token}{Access token retrieved from `getToken()`}

\item{agency_ids}{Optional vector of agency IDs to filter by}
}
\value{
An sf object containing spatial features data
}
\description{
Retrieves VectorSurv spatial features data with optional agency filtering
}
\examples{
\dontrun{token=getToken()
spatial = getSpatialFeatures(token)}
}
