% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpeciesMap.R
\name{plotSpeciesMap}
\alias{plotSpeciesMap}
\title{Plot Species Map}
\usage{
plotSpeciesMap(
  token,
  target_year,
  target_species = NULL,
  trap_type = NULL,
  agency_ids = NULL,
  time_period = NULL,
  interval = "Month",
  basemap = "Topographic",
  show_heatmap = FALSE,
  output_file = NULL,
  output_format = "leaflet",
  width = 1200,
  height = 800
)
}
\arguments{
\item{token}{x}

\item{target_year}{x}

\item{target_species}{x}

\item{trap_type}{x}
}
\description{
Plot map of mosquito species
}
